[IEE-22298] Some fixes and improvements

+++ dcm4chee-arr-cdi-query/src/main/java/org/dcm4chee/arr/cdi/query/fhir/FhirQueryDecorator.java
@@ -226,7 +226,7 @@ public class FhirQueryDecorator extends AbstractAuditRecordQueryDecorator
 		return this;
 	}	
 	
-	public FhirQueryDecorator setMaxResults( Integer maxResults )
+	public FhirQueryDecorator setMaxResults( Long maxResults )
 	{
 		super.setMaxResults( maxResults );
 		return this;
