+++ dcm4chee-arr-cdi-query/src/main/java/org/dcm4chee/arr/cdi/query/simple/SimpleQueryDecorator.java
@@ -149,7 +149,7 @@ public class SimpleQueryDecorator extends AbstractAuditRecordQueryDecorator
 		return this;
 	}
 
-	public SimpleQueryDecorator setMaxResults( Integer maxResults )
+	public SimpleQueryDecorator setMaxResults( Long maxResults )
 	{
 		super.setMaxResults( maxResults );
 		return this;
